v2

+++ cbr-cur/src/main/java/ru/tbank/json/CurrencyConverterResponse.java
@@ -7,6 +7,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -21,5 +23,5 @@ public class CurrencyConverterResponse {
     private String toCurrency;
 
     @JsonProperty("convertedAmount")
-    private Double convertedAmount;
+    private BigDecimal convertedAmount;
 }
